+++ content_service/src/main/java/io/urdego/content_service/api/user/controller/external/response/UserContentListAndCursorIdxResponse.java
@@ -1,4 +1,4 @@
-package io.urdego.content_service.api.user.controller.response;
+package io.urdego.content_service.api.user.controller.external.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
