+++ content_service/src/main/java/io/urdego/content_service/api/user/service/UserContentService.java
@@ -1,7 +1,7 @@
 package io.urdego.content_service.api.user.service;
 
-import io.urdego.content_service.api.user.controller.request.ContentUploadRequest;
-import io.urdego.content_service.api.user.controller.response.UserContentListAndCursorIdxResponse;
+import io.urdego.content_service.api.user.controller.external.request.ContentUploadRequest;
+import io.urdego.content_service.api.user.controller.external.response.UserContentListAndCursorIdxResponse;
 
 import org.springframework.web.multipart.MultipartFile;
